Fixes build error on MSVC, use of sqrt was ambiguous. Refs #1572
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@12804 3db68df2-c116-0410-a063-a993310a9797

+++ source/renderer/WaterManager.cpp
@@ -399,7 +399,7 @@ void WaterManager::CreateSuperfancyInfo()
 				}
 				
 			}
-			distanceToShore = sqrt(distanceToShore);
+			distanceToShore = sqrt((float)distanceToShore);
 			// Compute the normals
 			// Also create the waves quad.
 			CVector3D normal;
